make kernel bindings mutually exclusive
When the crate is built with --all-features we have a bunch of errors
because we are not excluding the 4.14 and 4.20 bindings. Whenever
all features is specified now, the 4.20 bindings are going to be
used. This change is required for switching to rust-vmm-ci where
the tests are run using --all-features flag.

Signed-off-by: Andreea Florescu <fandree@amazon.com>