arm64: add KVM_REG_ARM64_SVE_VLS const
For using SVE scenario, KVM_REG_ARM64_SVE_VLS is used by
set_one_reg/get_one_reg ioctls, which allows the set of vector lengths
supported by the vcpu to be discovered and configured by userspace. See
https://www.kernel.org/doc/html/latest/virt/kvm/api.html#kvm-set-one-reg

When using bindgen, this const is missing, so add it according to the
definition in linux.

Signed-off-by: xuejun-xj <jiyunxue@linux.alibaba.com>