Avoid temporary Vec<u8> allocation during deserialization
Instead of reusing `Vec::<u8>::deserialize` to defer the
deserialization, implement a Visitor that can construct `$typ` directly
from the deserialization byte stream. This eliminates a temporary
allocation where the input byte stream is first copied into a `Vec<u8>`
and later copied to the stack to construct `$typ` via `transmute!`.

This saves a few microseconds (up to 1ms on specific hardware) when
deserializing microVM snapshots.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>